Extract adjudication data for
